# Implement lazy record parsing (issue #30)