# Implement Lazy Record Parsing

## Overview

This PR implements lazy record parsing as described in issue #30, significantly improving query performance when selecting a subset of columns.

## What Changed

- Added `LazyRecord` struct that parses column values on-demand instead of eagerly
- Updated `BTreeCursor` to use lazy records throughout the query execution pipeline  
- Modified VDBE opcodes to work with lazy records
- Implemented zero-copy design with efficient caching

## Performance Results

Benchmarked on M3 MacBook Air with 1000 rows from the `users` table:

```
SELECT id FROM users LIMIT 1000:              47.4% faster (121.72 → 64.05 µs)
SELECT id, email FROM users LIMIT 1000:       39.5% faster (153.89 → 93.13 µs)  
SELECT id, email, first_name FROM ...:        34.6% faster (180.69 → 118.26 µs)
SELECT id, email, first_name, last_name ...:  28.6% faster (220.74 → 157.61 µs)
SELECT * FROM users LIMIT 1000:               17.3% faster (343.84 → 284.23 µs)
```

## Key Benefits

- **Reduced CPU usage**: Only parse columns that are actually accessed
- **Lower memory allocation**: Zero-copy design avoids unnecessary allocations  
- **Consistent improvement**: Even SELECT * queries benefit from streaming comparisons
- **No regressions**: All 537 tests pass, full SQLite compatibility maintained

## Implementation Highlights

1. **Smart caching**: Parsed values are cached to avoid re-parsing
2. **O(1) column access**: Offset table enables direct access to any column
3. **Streaming comparisons**: Compare values without materializing entire records
4. **Overflow handling**: Properly handles records spanning multiple pages

## Testing

- All existing tests pass
- Added unit tests for lazy parsing
- Comprehensive benchmark suite  
- No clippy warnings
- Code properly formatted

Fixes #30